refactor: autoVarName 함수 중복 제거 (#2)
서버/웹 양쪽에 흩어진 autoVarName, autoLlmVarName, autoCodeVarName 로직을
각각 autoVarNames.js 단일 파일로 추출. 기존 파일들은 import로 대체.
